+++ src/menu.tcl
@@ -170,7 +170,7 @@ proc menu::FillMenu {} {
 
   ## ________________________ Tools _________________________ ##
   set m [set al(TOOLS) $al(WIN).menu.tool]
-  $m add command -label $al(MC,run) -command alited::tool::_run -accelerator $al(acc_3)
+  $m add command -label [msgcat::mc Run...] -command alited::tool::RunMode -accelerator $al(acc_3)
   $m add command -label e_menu -command {alited::tool::e_menu o=0} -accelerator $al(acc_2)
 
     ### ________________________ Runs _________________________ ###